Refactor temporally component commerce-account
