--- blocks/commerce-account/api/customer/getCustomerCart.js
@@ -1,25 +1,0 @@
-import GqlClient from '../shared/GqlClient.js';
-
-const getCustomerCart = async (token) => {
-  const query = `
-    {
-      customerCart {
-        id
-        items {
-          id
-          product {
-            name
-            sku
-          }
-          quantity
-        }
-      }
-    }
-`;
-
-  const data = await GqlClient().fetchData(query, token);
-
-  return data;
-};
-
-export default getCustomerCart;
