--- blocks/commerce-account/api/customer/index.js
@@ -1,5 +1,0 @@
-export { default as getCustomer } from './getCustomer.js';
-export { default as getCustomerToken } from './getCustomerToken.js';
-export { default as getCustomerCart } from './getCustomerCart.js';
-export { default as getCustomerOrders } from './getCustomerOrders.js';
-export { default as getCustomerOrderById } from './getCustomerOrderById.js';
